[Android][Shared Core] Add create internal script
This script is used to create xwalk core internal from
current xwalk core.
The reason to scripten this process is to make it easy
for rebasing.
The whole feature for shared xwalk core is so complex that
it will take amount of time to review. The rebase for
moving core to core_internal will be a nightmare then.